requirements.txt
pyyaml==6.0.2
torch==2.8.0
torchao==0.13.0
torchvision==0.23.0
tensorflow==2.20.0
onnxscript==0.5.0
onnxruntime-gpu
tensorboard==2.20.0
torch_tb_profiler
fastapi
uvicorn
tensorboard-plugin-profile==2.20.6
torch-pruning